fix(backend/catalog): align domain, infra tests, and integration tests with current schema
Adds Metadata to Category and UpdateCategoryRequest, adds StripeProductID to
ProductRes, fixes a missing return statement in GetActiveImage, and removes
stale Metadata assertions from tests that no longer exercise that path. Also
fixes int vs int64 mismatch in price test and checks errors on HTTP responses.

+++ backend/internal/catalog/infrastructure/stripe/price/create_price_test.go
@@ -94,7 +94,7 @@ func TestCreatePrice(t *testing.T) {
 	t.Run("create price with different currencies", func(t *testing.T) {
 		testCases := []struct {
 			currency string
-			amount   int
+			amount   int64
 		}{
 			{"EUR", 2500},
 			{"GBP", 2000},
